Restrukturering Flyttet diverse filer til /components og /containers. Merget Topplinje og Header-komponent til en.
