+++ src/containers/Arbeidsforhold.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import {hentSaksopplysninger, PersonSelector, ArbeidsforholdSelector} from './ducks/saksopplysninger';
+import {hentSaksopplysninger, PersonSelector, ArbeidsforholdSelector} from '../ducks/saksopplysninger';
 import {Panel} from 'nav-frontend-paneler';
 import './arbeidsforhold.css';
 import { Systemtittel } from "nav-frontend-typografi";
