+++ src/containers/arbeidsforholdDetalj.less
